Fix size_t to int conversions warnings, and main
Fixed a lot of loops where used converted between signed int and size_t

Also fixed main function.   Needs further work to implement stochastic options.

+++ src/basicTrellis.h
@@ -62,7 +62,7 @@ namespace StochHMM{
         void initBackward();
         
         
-        double getEndingTransition(int);
+        double getEndingTransition(size_t);
         
         double getTransition();
